test: further split coordinator update tests
